[backend] reorganize services and schemas
