+++ backend/schemas/__init__.py
@@ -1,0 +1,5 @@
+from .profile import ProfileSchema
+from .reminders import ReminderSchema
+from .timezone import TimezoneSchema
+
+__all__ = ["ProfileSchema", "ReminderSchema", "TimezoneSchema"]
