+++ backend/schemas/profile.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+class ProfileSchema(BaseModel):
+    telegram_id: int
+    icr: float
+    cf: float
+    target: float
+    low: float
+    high: float
