+++ backend/schemas/reminders.py
@@ -1,20 +1,6 @@
 from pydantic import BaseModel
 
 
-class TimezoneSchema(BaseModel):
-    telegram_id: int
-    tz: str
-
-
-class ProfileSchema(BaseModel):
-    telegram_id: int
-    icr: float
-    cf: float
-    target: float
-    low: float
-    high: float
-
-
 class ReminderSchema(BaseModel):
     telegram_id: int
     id: int | None = None
