+++ backend/schemas/timezone.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class TimezoneSchema(BaseModel):
+    telegram_id: int
+    tz: str
