+++ backend/services/__init__.py
@@ -1,0 +1,12 @@
+from diabetes.services.db import init_db
+
+from .profile import save_profile, set_timezone
+from .reminders import list_reminders, save_reminder
+
+__all__ = [
+    "init_db",
+    "set_timezone",
+    "save_profile",
+    "list_reminders",
+    "save_reminder",
+]
